Fixed code most seems to work now
But still having problems with PWR query and Velocity. The velocity query is generating values but when I increase the power the velocity is not increasing? not exactly certain why its not working as expected. Will work on moreu tomorrow. The PWM query returns 0 but should return the last PWM set values?