Fixed ENPWR and other enhancements
Had to use static variables to save pwm settings. Also added bunch of debug to ENVEL but it just shows that the encoders don't work well. and added new DEBUG command 254 which toggles serial prints of command type and other debug prints. DEBUG is initialized to TRUE but with 254 one can turn it off and not see any serial prints.